refactor: move processLevelUp to its own file
