+++ src/game-logic/reducers/index.test.js
@@ -4,7 +4,6 @@ import {
   ACHIEVEMENT_COMPLETED,
   COW_ATTRITION_MESSAGE,
   FERTILIZERS_PRODUCED,
-  LEVEL_GAINED_NOTIFICATION,
   LOAN_INCREASED,
   LOAN_PAYOFF,
   MILKS_PRODUCED,
@@ -31,9 +30,8 @@ import {
 } from '../../constants'
 import { huggingMachine, sampleCropItem1 } from '../../data/items'
 import { sampleRecipe1 } from '../../data/recipes'
-import { genders, standardCowColors, toolLevel } from '../../enums'
+import { genders, standardCowColors } from '../../enums'
 import {
-  farmProductSalesVolumeNeededForLevel,
   generateCow,
   getCostOfNextStorageExpansion,
   getCowFertilizerItem,
@@ -1277,108 +1275,6 @@ describe('sellItem', () => {
         })
       })
     })
-  })
-})
-
-describe('processLevelUp', () => {
-  test('shows notifications for each level gained in the sale', () => {
-    jest.resetModules()
-    jest.mock('../../data/levels', () => ({
-      levels: [
-        {
-          id: 1,
-          unlocksShopItem: 'sample-crop-seeds-1',
-        },
-      ],
-      itemUnlockLevels: {},
-    }))
-    const { todaysNotifications } = jest.requireActual('./').processLevelUp(
-      {
-        inventory: [],
-        itemsSold: {
-          'sample-crop-1': farmProductSalesVolumeNeededForLevel(3),
-        },
-        todaysNotifications: [],
-      },
-      1
-    )
-
-    expect(todaysNotifications).toEqual([
-      {
-        message: LEVEL_GAINED_NOTIFICATION`${3}${{ name: '' }}`,
-        severity: 'success',
-      },
-      {
-        message: LEVEL_GAINED_NOTIFICATION`${2}${{ name: '' }}`,
-        severity: 'success',
-      },
-    ])
-  })
-
-  test('when sprinkler is selected when it gets a level up boost, hoveredPlotRangeSize increase', () => {
-    jest.resetModules()
-    jest.mock('../../data/levels', () => ({
-      levels: [
-        {
-          id: 0,
-        },
-        {
-          id: 1,
-        },
-        {
-          id: 2,
-          increasesSprinklerRange: true,
-        },
-      ],
-      itemUnlockLevels: {},
-    }))
-    jest.mock('../../constants', () => ({
-      INITIAL_SPRINKLER_RANGE: 1,
-      SPRINKLER_ITEM_ID: 'sprinkler',
-    }))
-
-    const { hoveredPlotRangeSize } = jest.requireActual('./').processLevelUp(
-      {
-        hoveredPlotRangeSize: 1,
-        itemsSold: {
-          'sample-crop-1': farmProductSalesVolumeNeededForLevel(2),
-        },
-        selectedItemId: 'sprinkler',
-        todaysNotifications: [],
-      },
-      1
-    )
-
-    expect(hoveredPlotRangeSize).toEqual(2)
-  })
-
-  test('unlocksTool reward makes tool become available', () => {
-    jest.resetModules()
-    jest.mock('../../data/levels', () => ({
-      levels: [
-        {
-          id: 0,
-        },
-        {
-          id: 1,
-          unlocksTool: 'SHOVEL',
-        },
-      ],
-      itemUnlockLevels: {},
-    }))
-    const newState = jest.requireActual('./').processLevelUp(
-      {
-        itemsSold: {},
-        inventory: [],
-        todaysNotifications: [],
-        toolLevels: {
-          SHOVEL: toolLevel.UNAVAILABLE,
-        },
-      },
-      0
-    )
-
-    expect(newState.toolLevels['SHOVEL']).toEqual(toolLevel.DEFAULT)
   })
 })
